+++ src/aux-server/aux-web/aux-player/PlayerApp/PlayerApp.vue
@@ -47,7 +47,7 @@
                     </router-link>
                     <md-list-item @click="addSimulation()">
                         <md-icon>cloud</md-icon>
-                        <span class="md-list-item-text">Subscribe to Channel</span>
+                        <span class="md-list-item-text">Add Channel</span>
                     </md-list-item>
                     <md-list-item
                         v-for="simulation in simulations"
@@ -82,6 +82,8 @@
                             <md-tooltip md-direction="bottom">{{ versionTooltip }}</md-tooltip>
                         </span>
                     </md-list-item>
+
+                    <tagline></tagline>
                 </md-list>
             </md-drawer>